Make conference.behaviour.js work for erc20 conference tests
